pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-review-agent"
version = "1.0.0"
description = "AI-powered Pull Request Review Agent for multiple git servers"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Hackathon Team", email = "team@example.com"}
]
keywords = ["git", "pull-request", "review", "ai", "code-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "requests>=2.31.0",
    "click>=8.1.7",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "GitPython>=3.1.40",
    "pygithub>=1.59.1",
    "python-gitlab>=4.2.0",
    "pylint>=3.0.3",
    "flake8>=6.1.0",
    "black>=23.11.0",
    "bandit>=1.7.5",
    "openai>=1.3.0",
    "pandas>=2.1.3",
    "numpy>=1.25.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "responses>=0.24.1",
    "sphinx>=7.2.6",
]

[project.scripts]
pr-review = "pr_review_agent.cli:main"

[project.urls]
Homepage = "https://github.com/team/pr-review-agent"
Repository = "https://github.com/team/pr-review-agent.git"
Issues = "https://github.com/team/pr-review-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["pr_review_agent*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=pr_review_agent --cov-report=term-missing"
